Remove monitored: true, set up CLI and config
This removest the `monitored: true` setting in the `sourcify-chains.ts`
of the sourcify server.
As well as all other places it appears.
This way the monitor settings will be isolated from the server.

Instead a monitor can be started on certain chains with `chains.json`
in the project root or by passing
`--configPath`.

Also possible to modify the default configs with `--configPath`.